ci: generate `clas-stringspinner` documentation for the highest `clas12-mcgen` tag
The usage guide for `clas-stringspinner` is auto generated. It's
possible its documentation will be ahead of the latest `clas12-mcgen`
version's `clas-stringspinner`, thus confusing OSG portal users.

This PR starts a CI config for this repo, and generates
`clas-stringspinner` documentation based on the highest `clas12-mcgen`
tag (which is assumed to be available on OSG). The link from the OSG
portal should point to this documentation.